npus-aiml-agentic-automation/microsoft_agent_framework/requirements.txt
agent-framework
databricks-sdk 
fastapi 
uvicorn 
python-dotenv
langchain
langchain-chroma